+++ src/pages/wallet/AssetActions.tsx
@@ -3,6 +3,7 @@ import { flatten, uniq } from "ramda"
 import MonetizationOnOutlinedIcon from "@mui/icons-material/MonetizationOnOutlined"
 import ShortcutOutlinedIcon from "@mui/icons-material/ShortcutOutlined"
 import RestartAltIcon from "@mui/icons-material/RestartAlt"
+import OpenInNewIcon from "@mui/icons-material/OpenInNew"
 import { isDenomTerraNative } from "@terra.kitchen/utils"
 import { has } from "utils/num"
 import { useIsClassic } from "data/query"
@@ -10,7 +11,7 @@ import { useNetworkName } from "data/wallet"
 import { useIsWalletEmpty } from "data/queries/bank"
 import { useCW20Pairs } from "data/Terra/TerraAssets"
 import { useTFMTokens } from "data/external/tfm"
-import { InternalButton, InternalLink } from "components/general"
+import { InternalButton, InternalLink, ExternalIconLink } from "components/general"
 import { ExtraActions } from "components/layout"
 import { ModalButton } from "components/feedback"
 import { ListGroup } from "components/display"
@@ -44,6 +45,15 @@ const AssetActions = ({ token, symbol, balance }: Props) => {
         </ModalButton>
       )}
 
+      {!isClassic && token.startsWith("ibc/") && (
+        <ExternalIconLink
+          icon={<OpenInNewIcon style={{ fontSize: 18 }} />}
+          href={`https://bridge.terra.money`}
+        >
+          {t("Bridge")}
+        </ExternalIconLink>
+      )}
+
       <InternalLink
         icon={<ShortcutOutlinedIcon style={{ fontSize: 18 }} />}
         to={`/send?token=${token}`}
@@ -57,9 +67,7 @@ const AssetActions = ({ token, symbol, balance }: Props) => {
           icon={<RestartAltIcon style={{ fontSize: 18 }} />}
           to="/swap"
           state={token}
-          disabled={
-            isWalletEmpty || !has(balance) || !getIsSwappableToken(token)
-          }
+          disabled={isWalletEmpty || !has(balance) || !getIsSwappableToken(token)}
         >
           {t("Swap")}
         </InternalLink>
